Split reducers into separate files
