+++ src/redux/reducers/accessToken.js
@@ -1,0 +1,17 @@
+import { LOGIN, LOGOUT } from "../actionTypes";
+
+const accessToken = (state = null, action) => {
+  switch (action.type) {
+    case LOGIN: {
+      const { accessToken } = action.payload;
+
+      return accessToken;
+    }
+    case LOGOUT:
+      return null;
+    default:
+      return state;
+  }
+}
+
+export default accessToken;
