+++ src/redux/reducers/index.js
@@ -1,0 +1,8 @@
+import { combineReducers } from "redux";
+import accessToken from './accessToken';
+import reminders from './reminders';
+
+export default combineReducers({
+  accessToken,
+  reminders
+});
